feat: campaign reward claim contract
